Fix bundle dependencies
Gyp actions are correct within themselves, change an
input causes a new output to be created, but it does
not seem that the action inputs and outputs are used
to chain the target dependencies.  This causes the
bundle pieces to be built out of order on Windows if
a previous build was done.